refactoring to the products and categories routes and and adding and editing products from the front end, also added a logger to log erros on the back-end

+++ backend/routes/products.routes.js
@@ -5,6 +5,7 @@ const router = express.Router();
 const mongoose = require("mongoose");
 const { count } = require("console");
 const multer = require("multer");
+const logger = require("../logger/logger");
 
 const FILE_TYPE_MAP = {
   "image/png": "png",
@@ -12,6 +13,7 @@ const FILE_TYPE_MAP = {
   "image/jpg": "jpg",
 };
 
+//Multer config
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
     const isFileValid = FILE_TYPE_MAP[file.mimetype];
@@ -31,25 +33,55 @@ const storage = multer.diskStorage({
 
 const uploadOptions = multer({ storage: storage });
 
+//Getting list of products
 router.get("/", async (req, res) => {
   let queryParams = {};
   if (req.query.categories) {
     queryParams = { category: req.query.categories.split(",") };
   }
-  const productList = await Product.find(queryParams);
+  try {
+    const productList = await Product.find(queryParams).populate("category");
 
-  if (!productList) {
-    res.status(500).send({ success: false });
+    if (!productList) {
+      logger.productsRoutesErrorLogger.log(
+        "error",
+        "Error while fetching products list"
+      );
+      return res.status(500).send({
+        success: false,
+        message: "Something went wrong, please try again later",
+      });
+    } else {
+      res.send(productList);
+    }
+  } catch (error) {
+    console.log(error);
+    logger.productsRoutesErrorLogger.log(
+      "server-error",
+      "Server error while fetching products list"
+    );
+    return res.status(500).send({ success: false, message: "Server error" });
   }
-
-  res.send(productList);
 });
 
+//Getting product by id
 router.get("/:id", async (req, res) => {
+  const id = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    logger.productsRoutesErrorLogger.log(
+      "error",
+      "Invalid mongo id provided when trying to find a product by id"
+    );
+    return res.status(400).send({ message: "Invalid category id provided." });
+  }
   try {
-    const product = await Product.findById(req.params.id).populate("category");
+    const product = await Product.findById(id).populate("category");
     if (!product) {
-      res.status(404).send({
+      logger.productsRoutesErrorLogger.log(
+        "error",
+        "Error while fetching product by id"
+      );
+      return res.status(404).send({
         success: false,
         message: "Product with this id does not exist!",
       });
@@ -57,6 +89,10 @@ router.get("/:id", async (req, res) => {
       res.send(product);
     }
   } catch (error) {
+    logger.productsRoutesErrorLogger.log(
+      "error",
+      "Error while fetching products list"
+    );
     return res.status(500).send({
       success: false,
       message: "Something went wrong, please try again later!",
@@ -64,60 +100,94 @@ router.get("/:id", async (req, res) => {
   }
 });
 
+//Adding a new product
 router.post("/", uploadOptions.single("image"), async (req, res) => {
-  const category = await Category.findById(req.body.category);
-
-  if (!category) {
-    return res.status(404).send("This category does not exist.");
-  }
-
-  const imageFile = req.file;
-  const fileName = imageFile.filename;
-  const basePath = `${req.protocol}://${req.get("host")}/public/uploads/`;
-
-  if (!imageFile)
-    return res
-      .status(400)
-      .send({ success: false, message: "No image file provided" });
-
-  let product = new Product({
-    name: req.body.name,
-    description: req.body.description,
-    richDescription: req.body.richDescription,
-    image: `${basePath}${fileName}`,
-    brand: req.body.brand,
-    price: req.body.price,
-    category: req.body.category,
-    countInStock: req.body.countInStock,
-    rating: req.body.rating,
-    numReviews: req.body.numReviews,
-    isFeatured: req.body.isFeatured,
-  });
-
-  product = await product.save();
-
-  if (!product) {
-    return res
-      .status(500)
-      .send("Product was not created! Please try again later.");
-  }
-  res.send(product);
-});
-
-router.put("/:id", uploadOptions.single("image"), async (req, res) => {
-  if (!mongoose.isValidObjectId(req.params.id))
-    return res
-      .status(404)
-      .send({ success: false, message: "The provided id is wrong!" });
   try {
     const category = await Category.findById(req.body.category);
-    if (!category)
+
+    if (!category) {
+      logger.productsRoutesErrorLogger.log(
+        "cad-request-error",
+        "Error find the category when adding a new product"
+      );
+      return res.status(404).send("This category does not exist.");
+    }
+
+    const imageFile = req.file;
+    if (!imageFile) {
+      logger.productsRoutesErrorLogger.log("error", "No image file provided");
+      return res
+        .status(400)
+        .send({ success: false, message: "No image file provided" });
+    }
+
+    const fileName = imageFile.filename;
+    const basePath = `${req.protocol}://${req.get("host")}/public/uploads/`;
+
+    let product = new Product({
+      name: req.body.name,
+      description: req.body.description,
+      richDescription: req.body.richDescription,
+      image: `${basePath}${fileName}`,
+      brand: req.body.brand,
+      price: req.body.price,
+      category: req.body.category,
+      countInStock: req.body.countInStock,
+      rating: req.body.rating,
+      numReviews: req.body.numReviews,
+      isFeatured: req.body.isFeatured,
+    });
+
+    product = await product.save();
+
+    if (!product) {
+      logger.productsRoutesErrorLogger.log(
+        "bad-request-error",
+        "Error while adding a new product"
+      );
+      return res
+        .status(500)
+        .send("Product was not created! Please try again later.");
+    }
+    return res.status(200).send(product);
+  } catch (error) {
+    logger.productsRoutesErrorLogger.log("server-error", "Server Error");
+    console.log(error);
+    return res.status(500).send({ error: error, message: "Server Error" });
+  }
+});
+
+//Updating a product
+router.put("/:id", uploadOptions.single("image"), async (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.id)) {
+    logger.productsRoutesErrorLogger.log(
+      "id-error",
+      "Invalid mongo id provided when trying to update a product by id"
+    );
+    return res
+      .status(404)
+      .send({ success: false, message: "Invalind mongoDB id" });
+  }
+  try {
+    const category = await Category.findById(req.body.category);
+    if (!category) {
+      logger.productsRoutesErrorLogger.log(
+        "error",
+        "Category not found while trying to update a product"
+      );
       return res
         .status(404)
         .send({ success: false, message: "Not valid Category" });
+    }
 
     const product = await Product.findById(req.params.id);
-    if (!product) return res.status(400).send("Invalid product");
+    if (!product) {
+      logger.productsRoutesErrorLogger.log(
+        "bad-request-error",
+        "Error updating the product"
+      );
+      return res.status(400).send("Invalid product");
+    }
 
     const imageFile = req.file;
     let imagePath;
@@ -148,14 +218,22 @@ router.put("/:id", uploadOptions.single("image"), async (req, res) => {
     );
 
     if (!updatedProduct) {
-      res.status(404).send({
+      logger.productsRoutesErrorLogger.log(
+        "bad-request-error",
+        "Error finding the updated product"
+      );
+      return res.status(404).send({
         success: false,
         message: "The product with this id does not exist!",
       });
     } else {
-      res.status(200).send(updatedProduct);
+      return res.status(200).send(updatedProduct);
     }
   } catch (error) {
+    logger.productsRoutesErrorLogger.log(
+      "server-error",
+      "Server error while updating a product"
+    );
     return res.status(500).send({
       success: false,
       message: "Something went wrong, try again later!",
@@ -168,40 +246,70 @@ router.put(
   "/gallery/images/:id",
   uploadOptions.array("images", 10),
   async (req, res) => {
-    if (!mongoose.isValidObjectId(req.params.id))
+    if (!mongoose.isValidObjectId(req.params.id)) {
+      logger.productsRoutesErrorLogger.log(
+        "id-error",
+        "Invalid mongo id provided when trying to update images of product by id"
+      );
       return res
         .status(404)
         .send({ success: false, message: "The provided id is wrong!" });
-
-    const imageFiles = req.files;
-    let imagesPaths = [];
-    const basePath = `${req.protocol}://${req.get("host")}/public/uploads/`;
-
-    if (imageFiles) {
-      imageFiles.map((file) => {
-        imagesPaths.push(`${basePath}${file.filename}`);
-      });
     }
+    try {
+      const imageFiles = req.files;
+      let imagesPaths = [];
+      const basePath = `${req.protocol}://${req.get("host")}/public/uploads/`;
 
-    const product = await Product.findByIdAndUpdate(
-      req.params.id,
-      {
-        images: imagesPaths,
-      },
-      { new: true }
-    );
-    if (!product) {
-      res.status(404).send({
+      if (imageFiles) {
+        imageFiles.map((file) => {
+          imagesPaths.push(`${basePath}${file.filename}`);
+        });
+      }
+
+      const product = await Product.findByIdAndUpdate(
+        req.params.id,
+        {
+          images: imagesPaths,
+        },
+        { new: true }
+      );
+      if (!product) {
+        logger.productsRoutesErrorLogger.log(
+          "bad-request-error",
+          "Error finding the product"
+        );
+        res.status(404).send({
+          success: false,
+          message: "The product with this id does not exist!",
+        });
+      } else {
+        return res.status(200).send(product);
+      }
+    } catch (error) {
+      logger.productsRoutesErrorLogger.log(
+        "server-error",
+        "Server error while updating a products images"
+      );
+      return res.status(500).send({
         success: false,
-        message: "The product with this id does not exist!",
+        message: "Something went wrong, try again later!",
+        error: error,
       });
-    } else {
-      res.status(200).send(product);
     }
   }
 );
 
 router.delete("/:id", async (req, res) => {
+  const id = req.params.id;
+  if ((!id, mongoose.isValidObjectId)) {
+    logger.productsRoutesErrorLogger.log(
+      "id-error",
+      "Invalid mongo id provided when trying to delete a product by id"
+    );
+    return res
+      .status(404)
+      .send({ success: false, message: "The provided id is wrong!" });
+  }
   try {
     const product = await Product.findByIdAndRemove(req.params.id);
     if (product) {
@@ -214,6 +322,10 @@ router.delete("/:id", async (req, res) => {
         .send({ success: false, message: "Product not found!" });
     }
   } catch (error) {
+    logger.productsRoutesErrorLogger.log(
+      "server-error",
+      "Server error while deleting a category"
+    );
     return res.status(500).send({
       success: false,
       message: "Something went wrong, please try again later...",
@@ -230,6 +342,10 @@ router.get("/get/count", async (req, res) => {
         .send({ success: true, totalProductCount: totalProductCount });
     }
   } catch (error) {
+    logger.productsRoutesErrorLogger.log(
+      "server-error",
+      "Error while trying to get the total products count"
+    );
     return res.status(500).send({
       success: false,
       message: "Something went wrong, please try again later!",
@@ -253,6 +369,10 @@ router.get("/get/featured/:count?", async (req, res) => {
         .send({ success: false, message: "There are no featured products!" });
     }
   } catch (error) {
+    logger.productsRoutesErrorLogger.log(
+      "server-error",
+      "Error while trying to get the total featured products count"
+    );
     return res.status(500).send({
       success: false,
       message: "Something went wrong, please try later!",
